+++ CHANGELOG.md
@@ -5,6 +5,14 @@ This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 
 ##### master
 
+**1.0.0-beta2**
+
+- NEW: Added whois.cdmon.com parser (GH-27). [Thanks @sfumanal]
+
+- FIXED: Fix for Record#respond_to?(:available?) (GH-28, GH-29, GH-30). Thanks [@marcandre]
+
+**1.0.0-beta1**
+
 Initial import from the `whois` library.
 
 - NEW: whois.dk-hostmaster.dk parser now recognizes throttled responses (whois/GH-382). [Thanks @troelskn]